pyproject.toml
[tool.poetry]
name = "marathon-training-app"
version = "0.1.0"
description = "Marathon Training App"
authors = ["Nuria nurimateosg@gmail.com"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
#streamlit = { version = "^1.25.0", python = ">=3.8,<3.9.7 || >3.9.7" }
#numpy = { version = "^1.25.2", python = ">=3.9" }
oauth2 = "^1.9.0.post1"
google-api-python-client = {version = "^2.99.0", python = ">=3.7"}
google-auth = { version = "^2.23.0", python = ">=3.7" }
google-auth-oauthlib = { version = "^1.1.0", python = ">=3.6" }
google-auth-httplib2 = "^0.1.1"
protobuf = {version = "^3.20.1", python = ">=3.7"}
altair = {version = "^4.2.2", python = ">=3.8"}
garminconnect = "0.1.55"
flask = { version = "^2.3.3", python = ">=3.8" }
requests = { version = "^2.31.0", python = ">=3.7" }
flask-cors = "^4.0.0"
openai = { version = "1.1.1", python = ">=3.7.1" }








